Revert "Merge branch 'WNP_Reunion' into user/vemancha/pushrequestunittests"
This reverts commit 6992a9d28b6c0a50944029a615006abb9a32533f, reversing
changes made to 6405c93f42935ad79f7a6870b2dd5380ce03e3da.

+++ installer/dev/install.cpp
@@ -103,7 +103,7 @@ namespace ProjectReunionInstaller {
     wil::com_ptr<IStream> CreateMemoryStream(const BYTE* data, size_t size)
     {
         wil::com_ptr<IStream> retval;
-        retval.attach(::SHCreateMemStream(data, static_cast<UINT>(size)));
+        retval.attach(::SHCreateMemStream(data, size));
         return retval;
     }
